# node-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Symptom Logger API",
    "version": "1.0.0",
    "description": "API documentation for the Symptom Logger API"
  },
  "servers": [
    {
      "url": "http://localhost:5200"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/provider": {
      "post": {
        "summary": "Create a new provider",
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "specialty": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "countryCode": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "Get all providers",
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "specialty": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "location": {
                        "type": "object",
                        "properties": {
                          "city": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string"
                          },
                          "countryCode": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/provider/{id}": {
      "get": {
        "summary": "Get a provider by ID",
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The provider ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "specialty": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "location": {
                      "type": "object",
                      "properties": {
                        "city": {
                          "type": "string"
                        },
                        "state": {
                          "type": "string"
                        },
                        "countryCode": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Provider not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing provider",
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The provider ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "specialization": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "countryCode": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Provider not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a provider by ID",
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The provider ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider deleted successfully"
          },
          "404": {
            "description": "Provider not found"
          }
        }
      }
    },
    "/api/report/user-symptom-log": {
      "get": {
        "summary": "Get all user symptom logs",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "required": false,
            "description": "The number of logs to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "required": false,
            "description": "The number of logs to skip"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "date",
                "city",
                "country",
                "age",
                "gender"
              ]
            },
            "required": false,
            "description": "The field to sort by"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "required": false,
            "description": "The order to sort by"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of user symptom logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "userId": {
                        "type": "string"
                      },
                      "symptom": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "city": {
                        "type": "string"
                      },
                      "country": {
                        "type": "string"
                      },
                      "age": {
                        "type": "integer"
                      },
                      "gender": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/user-symptom-log": {
      "post": {
        "summary": "Log a symptom for the user",
        "tags": [
          "UserSymptomLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symptoms": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "severity": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        "symptomStart": {
                          "type": "string",
                          "format": "date"
                        },
                        "symptomEnd": {
                          "type": "string",
                          "format": "date"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "countryCode": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Symptom log created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "Get all the symptoms logged by the user",
        "tags": [
          "UserSymptomLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of user symptom logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "symptoms": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "severity": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 10
                            },
                            "symptomStart": {
                              "type": "string",
                              "format": "date"
                            },
                            "symptomEnd": {
                              "type": "string",
                              "format": "date"
                            },
                            "description": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "location": {
                        "type": "object",
                        "properties": {
                          "city": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string"
                          },
                          "countryCode": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}